+++ src/view/svelte/ProgressView/ProgressFolder.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type WordStatisticsPlugin from "src/main";
 	import { WSEvents, WSFolderEvent } from "src/model/events";
-	import { WSFolder } from "src/model/folder";
+	import { RECORDING, WSFolder } from "src/model/folder";
 	import { FormatNumber, FormatWords } from "src/util";
 	import { onDestroy, onMount } from "svelte";
 	import ProgressBar from "./ProgressBar.svelte";
@@ -23,6 +23,9 @@
 	let title = folder.getTitle();
 	let name = folder.name;
 
+	let inherit: boolean;
+	let recording: boolean;
+
 	onMount(() => {
 		events.on(WSEvents.Folder.WordsChanged, onFolderUpdate, {
 			filter: null,
@@ -31,6 +34,7 @@
 		events.on(WSEvents.Folder.GoalSet, onFolderUpdate, { filter: null });
 		events.on(WSEvents.Folder.TitleSet, onFolderUpdate, { filter: null });
 		events.on(WSEvents.Setting.Recording, onFolderUpdate, { filter: null});
+		events.on(WSEvents.Folder.RecordingSet, onFolderUpdate, { filter: null });
 		updateAll();
 	});
 
@@ -42,6 +46,7 @@
 		events.off(WSEvents.Folder.GoalSet, onFolderUpdate, { filter: null });
 		events.off(WSEvents.Folder.TitleSet, onFolderUpdate, { filter: null });
 		events.off(WSEvents.Setting.Recording, onFolderUpdate, { filter: null});
+		events.off(WSEvents.Folder.RecordingSet, onFolderUpdate, { filter: null });
 	});
 
 	function onFolderUpdate(event: WSFolderEvent) {
@@ -66,15 +71,18 @@
 			countText = goal > 0 ? FormatNumber(count) : FormatWords(count);
 			label = goal > 0 ? countText + " / " + goalText : countText;
 			progress?.SetProgress(goal > 0 ? (count / goal) * 100 : 0);
+			inherit = folder.recording === RECORDING.INHERIT;
+			recording = folder.isRecording;
 		} else {
 			countText = "";
 			label = "";
 		}
+		folder = folder;
 	}
 </script>
 
 <div class="ws-progress-folder">
-	<h2><span class="record" class:recording={folder.isRecording}/>{title}</h2>
+	<h2><span class="record" class:recording={recording && !inherit} class:inherit={recording && inherit}/>{title}</h2>
 	{#if title != name}
 		<div class="ws-progress-folder-name">({name})</div>
 	{/if}